First step for cleaning up the file name assignment mess
Given that the average XS module has little to no Perl code, it's very unlikely
that there will be any sample allowing us to map package to file name, and the
code to do it makes aggregation more complicated than it needs to be.

Just generate a synthetic name for the file, based on the sub packge. It will
typically match one of the files on disk, which makes more sense than grouping
everything under "(unknown)", and allows us to use heuristics and other lies to
map the package to a .pm file on disk.